requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database Dependencies (PostgreSQL)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Vector Store (Chroma for conversation memory)
chromadb>=0.4.15
sentence-transformers>=2.2.0

# Existing Blackbox API integration
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Development Dependencies  
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0